RandomUserApi v0.5.0 RandomUserApi.CLI

Handle the command line interface for getting random users.

Functions

get_female()

Gets single random user with gender “female”. Return map with the female.

get_females(n)

Gets n number of users and all of them are females. Return list of maps with females.

get_male()

Gets single random user with gender “male”. Return map with the male.

get_males(n)

Gets n number of users and all of them are males. Return list of maps with males.

get_user()

Gets single random user. Return map with the user.

get_users(n)

Gets n number of users. Return list of maps with users.